W.I.P - Refatoração
commonFunctions.js em processo de separação das funções em arquivos individuais.
