+++ src/components/alunos/js/formDeleteCurso.js
@@ -6,7 +6,8 @@ const db = getFirestore(firebaseApp);
 const auth = getAuth(firebaseApp);
 //---------------------------------------------------------------//
 //Components
-import { insertElementHTML, readableRandomStringMaker, btnCloseForm, defaultEventsAfterSubmitForm } from "../../jsCommon/commonFunctions.js";
+import { btnCloseForm, defaultEventsAfterSubmitForm } from "../../jsCommon/formsFunc.js";
+import  insertElementHTML from "../../jsCommon/insertElementHTML.js";
 import { insertOptionsSelectCursos } from "./commonAlunos.js";
 import { insertViewTableAlunosHTML } from "./viewTableAlunos.js";
 import { addLogInfo } from "../../logData/js/logFunctions.js";
@@ -32,17 +33,51 @@ function eventsFormDeleteCurso(RA, alunoNome, RG) {
   });
 }
 
-function closeConfirmBox(e) {
-  let confirmBox = e.target.closest('.confirm_box')
-  parent = confirmBox.parentElement;
-  parent.removeChild(confirmBox);
-  if (parent.className.includes('block_screen')) {
-    let block_screen = parent;
-    let parent_block_screen = block_screen.parentElement;
-    parent_block_screen.removeChild(block_screen);
-  }
-};
-export function confirmBoxDeleteCurso(target, curso, callback) {
+function setInfoFields(RA, alunoNome, RG) {
+  document.querySelector("#aluno_ra").value = RA;
+  document.querySelector("#aluno_nome").value = alunoNome;
+  document.querySelector("#aluno_rg").value = RG;
+  insertOptionsSelectCursos(RA)
+}
+
+//Deleta o curso associado ao aluno.
+function submitformDeleteCurso(e) {
+  let form = document.querySelector('#form_delete_curso');;
+  let RA = form.querySelector('#aluno_ra').value;
+  let selectCurso = form.querySelector('#select_curso');
+  let curso = selectCurso.value;
+  let idContrato = form.querySelector('#select_curso').options[selectCurso.selectedIndex].dataset.id_contrato;
+  deleteDoc(doc(db, 'alunato', RA, 'cursos', curso))
+    
+    .then(() => {
+      setDoc(doc(db, "contratos", idContrato),
+        {
+          metadata: {
+            aluno_associado: 'pendente'
+          }
+        },
+        { merge: true }
+      );
+    })
+    .then(() => {
+      defaultEventsAfterSubmitForm("#form_delete_curso", "Curso deletado com sucesso!");
+    }).then(() => {
+      setTimeout(() => {
+        insertViewTableAlunosHTML();
+      }, 2000)
+    }).then(() => {
+      addLogInfo('log_alunato', 'curso_deletado', RA + '-' + curso);
+    })
+    .catch((error) => {
+      addLogInfo('log_alunato', 'error', RA, error);
+      console.error("Erro ao adicionar curso: ", error)
+    });
+
+}
+
+
+//Confirma se o usuário realmente quer deletar o curso.
+ function confirmBoxDeleteCurso(target, curso, callback) {
   let elementTarget = document.querySelector(target);
   let bgMsgBox = document.createElement('div');
   bgMsgBox.className = 'block_screen';
@@ -71,52 +106,15 @@ export function confirmBoxDeleteCurso(target, curso, callback) {
   elementTarget.appendChild(bgMsgBox);
 }
 
-function setInfoFields(RA, alunoNome, RG) {
-  document.querySelector("#aluno_ra").value = RA;
-  document.querySelector("#aluno_nome").value = alunoNome;
-  document.querySelector("#aluno_rg").value = RG;
-  insertOptionsSelectCursos(RA)
-}
+function closeConfirmBox(e) {
+  let confirmBox = e.target.closest('.confirm_box')
+  parent = confirmBox.parentElement;
+  parent.removeChild(confirmBox);
+  if (parent.className.includes('block_screen')) {
+    let block_screen = parent;
+    let parent_block_screen = block_screen.parentElement;
+    parent_block_screen.removeChild(block_screen);
+  }
+};
 
-function submitformDeleteCurso(e) {
-  let form = document.querySelector('#form_delete_curso');;
-  let RA = form.querySelector('#aluno_ra').value;
-  let selectCurso = form.querySelector('#select_curso');
-  let curso = selectCurso.value;
-  let idContrato = form.querySelector('#select_curso').options[selectCurso.selectedIndex].dataset.id_contrato;
-
-  deleteDoc(doc(db, 'alunato', RA, 'cursos', curso))
-    .then(() => {
-      let data = new Date();
-      let id = data.getFullYear() + '' + (data.getMonth() + 1) + '' + data.getDate() + '' + readableRandomStringMaker(5);
-      setDoc(doc(db, "log", 'log_alunato'), {
-        [id]: `Curso referente ao contrato ${idContrato} deletado do aluno ${RA} em ${new Date()} por ${auth.currentUser.email}`
-      },
-        { merge: true })
-    })
-    .then(() => {
-      setDoc(doc(db, "contratos", idContrato),
-        {
-          metadata: {
-            aluno_associado: 'pendente'
-          }
-        },
-        { merge: true }
-      );
-    })
-    .then(() => {
-      defaultEventsAfterSubmitForm("#form_delete_curso", "Curso deletado com sucesso!");
-    }).then(() => {
-      setTimeout(() => {
-        insertViewTableAlunosHTML();
-      }, 2000)
-    }).then(() => {
-      addLogInfo('log_alunato', 'curso_deletado', RA + '-' + curso);
-    })
-    .catch((error) => {
-      addLogInfo('log_alunato', 'error', RA, error);
-      console.error("Erro ao adicionar curso: ", error)
-    });
-
-}
-
+//------------------------------------------------------------------------
