+++ src/components/controleAula/js/formAddFeedbackBimestral.js
@@ -5,7 +5,9 @@ const { getFirestore, setDoc, doc } = require("firebase/firestore")
 const db = getFirestore(firebaseApp);
 //---------------------------------------------------------------//
 //Components
-import { removeBugExtraBgFormBLockScreen, insertElementHTML, btnCloseForm, defaultEventsAfterSubmitForm } from "../../jsCommon/commonFunctions.js";
+import { removeBugExtraBgFormBLockScreen} from "../../jsCommon/commonFunctions.js";
+import { btnCloseForm, defaultEventsAfterSubmitForm } from "../../jsCommon/formsFunc.js";
+import insertElementHTML from "../../jsCommon/insertElementHTML.js";
 import { insertOptionsInSelectAluno, insertOptionSelectCurso } from "./formAddAula.js";
 import { addLogInfo } from "../../logData/js/logFunctions.js";
 //---------------------------------------------------------------//
