+++ src/components/cursosInfo/js/resumoCursosInfo.js
@@ -5,7 +5,7 @@ const { getFirestore, getDocs, collection } = require("firebase/firestore")
 const db = getFirestore(firebaseApp);
 //---------------------------------------------------------------//
 //Components
-import { insertElementHTML } from "../../jsCommon/commonFunctions.js";
+import insertElementHTML from "../../jsCommon/insertElementHTML.js";
 //---------------------------------------------------------------//
 export function insertResumoCursosInfo() {
     insertElementHTML('#home_content', './components/cursosInfo/resumoCursosInfo.html', eventsResumoCursosInfo, null, true);
