+++ src/components/cursosInfo/js/viewTableCursosInfo.js
@@ -7,7 +7,8 @@ const { getAuth } = require("firebase/auth");
 const auth = getAuth(firebaseApp);
 //---------------------------------------------------------------//
 //Components
-import { insertElementHTML, confirmBoxDelete, readableRandomStringMaker } from "../../jsCommon/commonFunctions.js";
+import insertElementHTML from "../../jsCommon/insertElementHTML.js";
+import { confirmBoxDelete } from "../../jsCommon/confirmBoxFunc.js";
 import { insertFormAddCursosInfoHTML } from "./formAddCursoInfo.js";
 import { removeUnauthorizedElement } from "../../../appContent/adminContent/js/checkPermission.js";
 import { addLogInfo } from "../../logData/js/logFunctions.js"
@@ -128,16 +129,8 @@ function createTableCursosInfoHTML(cursosInfo) {
     return tbody;
 }
 
-function submitDeleteCursoInfo(idCurso, cursoNome) {
+function submitDeleteCursoInfo(idCurso) {
     deleteDoc(doc(db, 'cursos_info', idCurso))
-        .then(() => {
-            let data = new Date();
-            let id = data.getFullYear() + '' + (data.getMonth() + 1) + '' + data.getDate() + '' + readableRandomStringMaker(5);
-            setDoc(doc(db, "log", 'log_cursos_info'), {
-                [id]: `Curso Info '${idCurso}, ${cursoNome}' deletado em ${new Date()} por ${auth.currentUser.email}`
-            },
-                { merge: true })
-        })
         .then(() => {
             insertViewTableCursosInfoHTML();
         }).then(() => {
