+++ src/components/fluxoCaixa/js/formAddSaidaAvulsa.js
@@ -6,7 +6,8 @@ const { getFirestore, doc, setDoc } = require("firebase/firestore")
 const db = getFirestore(firebaseApp);
 //---------------------------------------------------------------//
 import { setCurrentDate, converteMesNumeroPorExtenso } from '../../jsCommon/dateFunc.js';
-import { defaultEventsAfterSubmitFixedForm, insertElementHTML } from "../../jsCommon/commonFunctions.js";
+import { defaultEventsAfterSubmitFixedForm } from "../../jsCommon/formsFunc.js";
+import insertElementHTML from "../../jsCommon/insertElementHTML.js";
 import { addLogInfo } from '../../logData/js/logFunctions.js';
 //---------------------------------------------------------------//
 //Others libraries
