+++ src/components/logData/js/viewTableLogData.js
@@ -5,7 +5,7 @@ const { getFirestore, getDoc, doc } = require("firebase/firestore")
 const db = getFirestore(firebaseApp);
 //---------------------------------------------------------------//
 //Components
-import { insertElementHTML } from "../../jsCommon/commonFunctions.js";
+import insertElementHTML from "../../jsCommon/insertElementHTML.js";
 //---------------------------------------------------------------//
 
 export function insertViewTableDataLogHTML() {
