+++ src/components/login/js/index.js
@@ -9,7 +9,9 @@ const db = getFirestore(firebaseApp);
 const auth = getAuth(firebaseApp);
 //---------------------------------------------------------------//
 //Componentes
-import { importHTMLWithScript, insertElementHTML } from "../../jsCommon/commonFunctions.js";
+import importHTMLWithScript from "../../jsCommon/importHTMLWithScript.js";
+import insertElementHTML from "../../jsCommon/insertElementHTML.js";
+
 //---------------------------------------------------------------//
 
 export function onload() {
@@ -30,14 +32,11 @@ export function onload() {
         getDoc(doc(db, "metadata", "last_version")).then((res) => {
           if (res.data().version !== appVersion) {
             //Cerrega página de updateVersion
-            // importHTMLWithScript('#page_content', "../src/appContent/adminContent/updateAppVersion.html", "../../appContent/adminContent/js/index.js", eventsUpdateAppVersion);
             insertElementHTML('#app', './appContent/adminContent/updateAppVersion.html', () => {
               eventsUpdateAppVersion(res.data(), appVersion)
             }, null, true);
           } else {
             importHTMLWithScript('#app', './appContent/adminContent/index.html', "../../appContent/adminContent/js/index.js");
-            //Cerrega página principal da home
-            // importHTMLWithScript('#page_content', "./components/home/index.html", "../home/js/index.js");
           };
         })
         // ...
